src/iot_streamer/requirements.txt
python-dateutil==2.8.1
google-cloud-monitoring==0.34.0
